+++ src/nvim/move.c
@@ -794,7 +794,7 @@ void curs_columns(win_T *wp, int may_scroll)
       // column
       char_u *const sbr = get_showbreak_value(wp);
       if (*sbr && *get_cursor_pos_ptr() == NUL
-          && wp->w_wcol == (int)vim_strsize(sbr)) {
+          && wp->w_wcol == vim_strsize(sbr)) {
         wp->w_wcol = 0;
       }
     }
